requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.5
python-dotenv==1.0.1
pydantic==2.9.2